Added a link to `libstdc++` before we link to `libc++`.
This is a hack to get around a problem with plug-ins that have a flat
namespace and need the `libstdc++` symbols to be resolved first.
While there's no guarantee that this hack will always work, it has been
tested in 10.7 and 10.8, and the underlying issue was fixed by Apple
in 10.9, so I'm as comfortable as I can be with this solution.

fixes #1